Refactor: ts마이그레이션 (ts임포트 관련 설정)
