--- src/common/constants/index.js
@@ -1,2 +1,0 @@
-export * from "./httpStatus.js";
-export * from "./errorMessage.js";
