+++ src/config/passport.ts
@@ -2,8 +2,8 @@ import passport from "passport";
 import {
   accessTokenStrategy,
   refreshTokenStrategy,
-} from "../middlewares/passport/jwt.strategy.js";
-import googleStrategy from "../middlewares/passport/google.strategy.js";
+} from "../middlewares/passport/jwt.strategy";
+import googleStrategy from "../middlewares/passport/google.strategy";
 
 passport.use("access-token", accessTokenStrategy);
 passport.use("refresh-token", refreshTokenStrategy);
